pyproject.toml
[tool.poetry]
name = "moneyforward-paid-service"
version = "0.1.0"
description = ""
authors = ["hikaelis <sileakih@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.24.0"
python-dotenv = "^1.0.1"
webdriver-manager = "^4.0.2"
flake8 = "^7.1.1"
black = "^24.8.0"
isort = "^5.13.2"
line-bot-sdk = "^3.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 127

[tool.isort]
profile = "black"